Move isStale into the shared library; add a comment
